Migrate `horizontal-multivec.test.js`
